Refactor Phase 3: split config_discovery into discovery/ submodules
Splits the monolithic ConfigDiscovery class into focused collaborators
under src/connect_migrate/discovery/:

  worker_rest_client.py        WorkerRestClient: REST calls to Connect
                               worker (is_alive, get_json,
                               get_connector_configs, get_connector_statuses)
  local_file_loader.py         LocalFileLoader: parses local config files
                               (worker URLs, key=value worker configs)
  sensitive_data_redactor.py   SensitiveDataRedactor: redaction logic
  config_discovery.py          Thin back-compat facade preserving the
                               external surface used by the rest of the
                               codebase (constructor kwargs, FM_CONFIGS_DIR
                               class const, discover_and_save, and the three
                               static helpers get_json_from_url,
                               get_connector_configs_from_worker,
                               get_connector_statuses_from_worker).

The ~120-entry static sensitive-key list and the patterns list are lifted
into src/connect_migrate/constants/sensitive_keys.py as module-level data
(STATIC_SENSITIVE_CONFIG_KEYS, SENSITIVE_KEY_PATTERNS).

Import sites updated: connector_comparator, migration/offset_manager,
cli/discover_cli, cli/migrate_cli. No behavior change; --help unchanged;
back-compat surface verified.
