+++ src/connect_migrate/cli/migrate_cli.py
@@ -11,7 +11,7 @@ import logging
 
 from connect_migrate.migration.offset_manager import OffsetManager
 from connect_migrate.connector_comparator import ConnectorComparator
-from connect_migrate.config_discovery import ConfigDiscovery
+from connect_migrate.discovery.config_discovery import ConfigDiscovery
 from connect_migrate.utils.logging_setup import setup_logging
