+++ src/connect_migrate/constants/sensitive_keys.py
@@ -1,0 +1,128 @@
+"""Static data: keys that should always be redacted from connector configs."""
+
+from typing import List
+
+
+STATIC_SENSITIVE_CONFIG_KEYS: List[str] = [
+    "activemq.password",
+    "appdynamics.proxy.password",
+    "aws.access.key.id",
+    "aws.dynamodb.proxy.password",
+    "aws.lambda.proxy.password",
+    "aws.redshift.password",
+    "aws.secret.access.key",
+    "aws.secret.key.id",
+    "azblob.account.key",
+    "azblob.proxy.password",
+    "azblob.sas.token",
+    "azure.datalake.client.key",
+    "azure.datalake.gen2.client.key",
+    "azure.datalake.gen2.sas.key",
+    "azure.eventhubs.sas.key",
+    "azure.search.api.key",
+    "azure.servicebus.connection.string",
+    "azure.servicebus.sas.key",
+    "azure.sql.dw.password",
+    "basic.auth.user.info",
+    "bearer.token",
+    "cassandra.password",
+    "cassandra.ssl.truststore.password",
+    "confluent.topic.ssl.key.password",
+    "confluent.topic.ssl.keystore.password",
+    "confluent.topic.ssl.truststore.password",
+    "connection.password",
+    "database.password",
+    "delta.lake.token",
+    "diode.encryption.password",
+    "diode.encryption.salt",
+    "elastic.https.ssl.key.password",
+    "elastic.https.ssl.keystore.password",
+    "elastic.https.ssl.truststore.password",
+    "ftps.password",
+    "ftps.ssl.key.password",
+    "ftps.ssl.keystore.password",
+    "ftps.ssl.truststore.password",
+    "function.key",
+    "gcf.credentials.json",
+    "gcp.bigtable.credentials.json",
+    "gcp.dataproc.credentials.json",
+    "gcp.firebase.credentials.json",
+    "gcp.pubsub.credentials.json",
+    "gcp.spanner.credentials.json",
+    "gcp.spanner.proxy.password",
+    "gcs.credentials.json",
+    "gemfire.password",
+    "http.proxy.password",
+    "https.ssl.key.password",
+    "https.ssl.keystore.password",
+    "https.ssl.truststore.password",
+    "impala.ldap.password",
+    "influxdb.password",
+    "java.naming.security.credentials",
+    "kinesis.proxy.password",
+    "mongodb.password",
+    "mq.password",
+    "mq.tls.key.password",
+    "mq.tls.keystore.password",
+    "mq.tls.truststore.password",
+    "mqtt.password",
+    "mqtt.ssl.key.password",
+    "mqtt.ssl.key.store.password",
+    "mqtt.ssl.trust.store.password",
+    "oauth2.client.secret",
+    "oauth2.jwt.keystore.password",
+    "pagerduty.api.key",
+    "pagerduty.proxy.password",
+    "password",
+    "prometheus.listener.basic.auth.password",
+    "prometheus.listener.ssl.key.password",
+    "prometheus.listener.ssl.keystore.password",
+    "proxy.password",
+    "rabbitmq.https.ssl.key.password",
+    "rabbitmq.https.ssl.keystore.password",
+    "rabbitmq.https.ssl.truststore.password",
+    "redis.password",
+    "redis.ssl.keystore.password",
+    "redis.ssl.truststore.password",
+    "s3.proxy.password",
+    "s3.sse.customer.key",
+    "salesforce.consumer.secret",
+    "salesforce.jwt.keystore.password",
+    "salesforce.password",
+    "salesforce.password.token",
+    "sasl.jaas.config",
+    "schema.registry.ssl.key.password",
+    "schema.registry.ssl.keystore.password",
+    "schema.registry.ssl.truststore.password",
+    "servicenow.password",
+    "solace.password",
+    "solace.ssl.keystore.password",
+    "solace.ssl.truststore.password",
+    "splunk.collector.authentication.tokens",
+    "splunk.hec.ssl.trust.store.password",
+    "splunk.hec.token",
+    "splunk.s2s.ssl.key.password",
+    "splunk.ssl.key.store.password",
+    "sqs.proxy.password",
+    "ssl.key.password",
+    "ssl.keystore.certificate.chain",
+    "ssl.keystore.key",
+    "ssl.keystore.password",
+    "ssl.truststore.certificates",
+    "ssl.truststore.password",
+    "staging.s3.access.key.id",
+    "staging.s3.secret.access.key",
+    "syslog.ssl.key.password",
+    "teradata.password",
+    "tibco.password",
+    "tls.passphrase",
+    "tls.private.key",
+    "tls.public.key",
+    "v3.$username.auth.password",
+    "v3.$username.privacy.password",
+    "vertica.password",
+    "zendesk.password",
+]
+
+
+SENSITIVE_KEY_PATTERNS: List[str] = ["password", "token", "secret", "credential"]
