+++ src/connect_migrate/migration/offset_manager.py
@@ -1,7 +1,7 @@
 import logging
 from typing import List, Optional, Dict, Any, Union
 import requests
-from connect_migrate.config_discovery import ConfigDiscovery
+from connect_migrate.discovery.config_discovery import ConfigDiscovery
 
 
 class OffsetManager:
